Stopping this development of the branch
It doesn't seem to produce as fast of a improvement or Graphic. the main reason why it is like this is due to the graphics scaling to Display. to fix this you have to use smooth scaling which is consuming on fps but seem to fix the graphic issue at the same speed as gfx